dodanie znaczkia czasu

+++ app/src/main/java/com/example/mapa_2/opis_nagrania.java
@@ -2,5 +2,5 @@ package com.example.mapa_2;
 
 public class opis_nagrania {
     double czas;
-    int nagranie;
+    long nagranie;
 }
